+++ packages/blueprint/jest.config.mjs
@@ -1,0 +1,7 @@
+import common from '../../jest.common.mjs';
+
+/** @type {import('@jest/types').Config.InitialOptions} */
+export default {
+  ...common,
+  displayName: 'blueprint',
+};