+++ website/src/pages/demo/_constants/types.ts
@@ -45,6 +45,7 @@ export type HttpsURL = `${'https'}://${string}`;
 export type StyleName =
   | 'default'
   | 'antd'
+  | 'blueprint'
   | 'bootstrap'
   | 'bulma'
   | 'chakra'